Run MonteCarlo on DDMA architecture
In this commit, I have written a script to run a Monte Carlo siulation
run to check and assess the performance of the DDMA scheme. The idea is
obtain curves showing the Angle error vs range SNR for different number
of DDMA ramps. IN each Monte carlo run, I randomize the number of
Dopplers in a given range from 3 to 9 so effectively the number of
components in the DOppler spectrum will be 12 to 36 (since 4 Txs). This
is to model the Steradian RADAR platforms which include eithere 12
TX/13Tx DDMA. So if there is a single Doppler in a given range bin, then
the Doppler spectrum would consist of 12 components. Since in my model,
the number of simultaneous Txs is 4 so I choose a minimum of 3 Dopplers
to obtain the 12 components as in Steradians RADAR plotform. Similarly,
the max Dopplers in a range bin can be assumed to be about 3. In which
case, Steradian's platform would consist of 12x3 = 36 Doppler
components. So, in order to achieve 36 Doppler components i my model, I
need 9 Dopplers in a range bin. On running the Monte carlo simulation, I
observed that the angle error is about 5 degrees even for very good
SNRs. I need to debug why this is happening.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>