Assign greater probability to lower number of Dopplers and lesser probability for higher number of Doppler bins
Previously, I was assigning uniform probabilities to the number of
Dopplers in a given range bin i.e. I was uniformly choosing the number
of Dopplers in a given range bin from 3 to 9 to mimic the sSteradian
MIMO system which has 12 Txs. So, 3 Dopplers in a given range bin was as
likely as having 9 Dopplers in the same range bin. But this is not the
trur scenario. We have a greater chance of having 1/2 Dopplers in a
given range than having 4/5 Dopplers. Hence I choose the number of
Dopplers in a given range bin based on a non-uniform distribution whihc
gives more weightage to lesser number of DOpplers bins and less
weightage to larger number of Doppler bins. numpy has an inbuilt package
'np.random.choice' which draws the desired number according to any given
distribution as specified by the user. With this change the Angle error
plots showed slight improvement but still it is not meeting the
expectations. In the next commit, I will model the MIMO array accordign
to the the Steradians 144v3 and 256v2 platforms which have 12 and 13
MIMO Txs respectively.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>