Model the DDMA scheme for Steradian SRIR144 and SRIR256 RADAR platforms
Till the previous commits, I was modelling the DDMA scheme with a toy
example with 4Txs separated by lambda/2 and 8 Txs each separated by
2lambda. In this commit, I have actually modelled the DDMA scheme for
the exact use case of Steradian's SRIR144 and SRIR256 platforms. SRIRR144 is a 3 IC platform with each IC having 4 Txs and 4 Rxs. Similarly, the SRIR256  is a 4 IC platform. However, the SRIR256 has only 13 usable Txs which effectively synthesizes a 13Tx x 16 Rx = 208 virtual MIMO elements. Also, the number of Az ULA elements for the 144 platform is 48 while the number of Az ULA elements
for the 256 platform is 74.

The MIMO phasor is computed as exp(1j*2pi/lamda [sin(theta)cos(phi) sin(theta)sin(phi) cos(theta)] [x y 0].T)
where theta is the angle in the plane of phi. x,y,z are the positions of each virtual MIMO channel. Since, all the
antennas are on the plane of the RADAR board, the z co-ordinate of the Antennas is always 0. I have written a function which generates the MIMO phasor for any of the selected platforms. This function is called by the 'DDMA_multiDoppler_vectorized.py' script to generate the DDMA MIMO signal. I have verified the script with both the 144 as well as 256 platform keys. It is working fine. This is the correct model to be used. In the next commits, I will incorporate this model for the single Doppler DDMA script as well as the monteCarlo script.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>