Modified script to cater to Steradian RADAR platforms
In this commit, I have modified the DDMA Monte Carlo script to cater to Steradian Imaging RADAR platforms namely SRIR144 and SRIR256. Since the number of Txs in these platforms are 12 and 13 respectively, I have changed the max Dopplers in  given range bin to 3. The Dopplers in a given range bin are drawn from a pmf where the likelihood of picking 1 Doppler is thrice the likelihood of picking 3 Dopplers. Similarly, the likelihood of picking 2 Dopplers is twice the likelihood of picking 3 Dopplers. Previously I was plotting only the standard deviation of the angle error but now I also plot the 98 percentile of the absolute angle error. This is because, on very rare occasions, we get a huge angle error and this can bias the std. Hence, the 98 percentile may be a good metric to judge the performance of the angle estimation in the DDMA scheme. This Monte Carlo run was taking a huge time on my personal laptop and also my laptop was restarting after sometime when this simulation was being run. So I ran this script on the lab Desktop machine which is even more powerful. It took 5 hours to complete the full run. The resulst are on expected lines.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>