Replace FFT with single point DFT
In this commit, I have replaced the 2048 point Doppler FFT with a single point
DFT to reduce compute. Previously, I was performing a 2048 point Doppler
FFT on the Doppler samples and this FFT was being performed for 16/12
Rxs (based on the platform 256 or 144) and then sampling the Doppler FFT
at the required Doppler bins and their modulation bins. This is a very
compute heavy operation. Instead, now I perform a DFT i.e a simple
correlation of the Doppler samples with the phasor generated using the
Doppler and its modulation bins. This way the compute falls drastically.
Just like we perform windowing prior to FFT, here also, prior to taking
the correlation/DFT, I window the Doppler samples. This is the major
change introduced in this commit. The other chages include some
comments, code clean up, etc.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>