Keystone transformation to fix range bin migration
Keystone transformation is an algorithm to fix the range bin migration in the detection segment where a single Tx transmits several chirps. In this commit, I have coded the Keystone transformation for an FMCW RADAR architecture. The reference paaper I have used is given as a link in the script. The keystone algorithm works on the raw ADC time samples and not on the range FFT data. The principle behind the keystone algorithm is that, if there is no range bin migration, then the phase variation across the chirp samples is the same no matter what ADC sample we choose. In other words, the phase variation across the ramps for the 100th ADC sample should be the same as the phase variation for the 1012 sample. This is because there will only be two terms namely the range term and the Doppler term and both of these are completely de-coupled.  The initial phase though will be different (because this manifests as range frequency) but the phase function/variation has to be the same. However, when there is a range bin migration across chirps (due to high velocity targets under high bandwidth conditions), then a range bin migration term is introduced which couples to both the range term as well as the Doppler term. Due to this, the Doppler phase function is no more similar for all the ADC samples. It is a function of which ADC time sample is chosen. The keystone tranformation removes this coupling/dependency term (range bin migration term) through a time dependent(ADC time dimension) chirp axis scaling. This is accomplished through interpolation along the chirp dimension (for each ADC time sample) and then resampling on a new chirp grid where each ADC sample has a different chirp resampling factor. The interpolating factor for an FMCW RADAR architecture turns out to be (fc/(fc+St)) where fc is the chirp centre frequency, S is the slope of the chirp.

For now, I have coded a single Doppler in a given range to test the Keystone algorithm that I had derived. It works perfectly fine! Currently, the keystone algorithm does not cater to velocities beyond +/-Fs/2. I need to understand why this is so. The reference paper has a separate section that deals with fixing the range bin migration when the velocity is greater than baseband velocity and has an "m" fold foldover to the baseband i.e. If v = v_b + mVs. In the subsequent commmits, I will generalize this script to cater to multiple Dopplers(all within +/-Fs/2) and with some noise. I will be interesting to see how well the interpolation works in the presence of noise.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>