Replaced first order spline interpolation with cubic spline interpolation
In this commit, I have replaced the linear(first order) spline
interpolation with the cubic spline interpolation. This is because,
cubic spline is a more smoother interpolation kernel as compared to
first order spline. Also, I have removed the smoothing factor 's' in the
arguments of the spline interpolation and made it generic.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>