Handle the crash in the function when there is only 1 local maxima
In this commit, I have handled the crash that was occuring in this
function when the number of local maxima in the signal is less than or
equal to 1. Now, if the number of local maxima in the passed signal is 1
or lesser, then I pick the 1st and last index of the signal to avoid the
crash. This way, it will not crash, but the two peaks given out in this
case will be wrong.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>