Plot Angle error curves for different FFT lengths
In this commit, I have plotted the angle error curves (accuracy curves)
for different FFT lengths like 256, 512,1024,2048,4096,8192. The signal
lenght is still 74 elemets but we compute 256 point FFT, 512 point FFT
and so on. This is to show how the angle accuracy imoproves with
increasing the FFT length. Also, I have randomized the angles from -50
to +50 degrees for each Monte Carlo run. Till the previous commit, I was
measuring the angle accuracy only at boresight i.e. the object angle was
set to 0 for all the Monte Carlo runs. However, the problem with this
simulation set up is that, once the SNR becomes high enough, then the
error will be 0 for all the runs and hence the std becomes 0 and hence
the log(std) becomes a large negative value and starts to violate the
Cramer Rao Lower Bound. To avoid this, I randomize the angles from -50
to +50 so that we get some qunatization in the FFT spectrum and with
this change, the graphs look fine. The lower FFT sizes follow the CRLB
curve upto a particular SNR point and beyond that, the curves saturate
because the error is now only limited by the FFT size/bin quantization
and not on the SNR.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>