Correct for aliased Dopplers using Keystone transformation
Till the last commit, I had developed the keystone transformation script to cater to a single target whose velocity was within the baseband limits set by the inter-chirp time. In this commit, I have extended the keystone algorithm to cater to a single target but whose velocity can be outside the baseband as set by the Doppler Interget Hypothesis/Doppler Ambiguity number. For this again, I had to go through the last section of the Keystone algorithm paper which addresses the issue of Doppler Ambiguity number. The paper showed the result for a pulsed RADAR whereas I have modified it to cater to FMCW RADAR architecture. Having an aliased Doppler causes an extra phase term exp(-1j*2pi*(fc/(fc+St))Mm), where fc is the chirp center frequency, S is the chirp slope, t is the ADC time, m is the chirp index and M is the Doppler Ambiguity number. We do not know what the M is. SO we hypothesize over different Ms say from -2 to +2 and see for which value of M the energy(post Range Dopler FFT) is maximum. We apply a conjugate of the above term to correct. We multiply this term post the Keystone transformation is applied. For now, since I know the hypothesis number, I have hard coded that. In the subsequency commits,  will sweep the Doppler Hypothesis number from say -2 to +2 and find out for which value of M, the energy is maximum and then select that number to get the true correction and hence the correct unaliased Doppler. I will derive the expressions in a separate pdf and add the link in the later commits.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>