Ensure phase shifter LUT is modulo 360 degrees
In this branch, I have made a minor code change to the DDMA scripts. The phase shifter LUT might not always be in the range from 0 to 360 degrees. It could be from say 160 degrees to 530 degrees and so when I take the closest value of the phase shifter code to say 0 degrees, I will end up picking 160. But this is wrong. I need to bring the phase shifter LUT modulo 360 degrees before searching for the closest code. For now, in my modeling, this is not very significant since I always generate the LUT from 0 to 360 degrees but in general, the phase shifter LUT might have a bias and might be unwrapped and could start from a high value and also go to a higher value.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>