Added docstring explaining the signal model and simulation conditions.
In this commit, I have added a docstring documenting the simulation
model. I have also modified the range spectrum plot by marking the
expected frequencies due to the FMCW + DAC model. Important point to
remember is that in this simulation model, in addition to the target
frequencies, we obtain the DAC + target frequencies on the positive side
of the spectrum and -(DAC-target frequencies) on the negative side of
the spectrum.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>