Fixed pointing and dropping of bits
This is my first commit in fixed point coding. I have added a folder named 'fixed_point_coding'. In this folder, I have added a first basic script which does a comparision of fixed and floating point operations. To begin with I have started with windowing of a signal. I compare the windowed signal with the fixed point implementation. Windowing is basically multiplication and I have taken the signal and the window to be 32 bit numbers and hence the product will be a 64 bit number. The idea is to get the output in fixed point 32 bit format and hence need to drop 32 bits post multiplication. In this case, I assume both the signal and window to be of 1Q31 format i.e. 1 integer bit which is also the signed bit and 31 fractional bits. Hence, the dynamic range possible with this format is a swing from +/-2^31. So when we multiply these two functions, the resultant signal will have a swing from +/-2^62. So, the effective fractional bits post multiplication is 62. To convert from 62 fractional bits to 31 fractional bits, we need to drop 31 fractional bits. The way we drop fractional bits is divide by 2**31 and then floor the resultant value. In hardware though this is done slightly done differently. If we have to drop 32 fractional bits, we first add 2**30 to the fixed point integer and then right shift by 31(since since division by a power of 2 can be achieved by right shifting operation). I have writtena separate function to drop fractional bits. Post this I compared the floating point windowed signal with the fixed point windowed signal. There seems to be sometimes an error of the order of 1e-3 for a signal of swing from +/-1. Need to check if this is tolerable or not and also whether this much error is expected or not. In the next commit, I will be adding fixed point implementation of matrix multiplication.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>