Replace computation of power of 2 by bit shift operations
In this commit, I have replaced the computation of a power of 2 with a
simple left shift operation of the binary number 1. More specifically,
if we need to compute 2^n, we can instead do 1 << n. Both are the same
operations. I have made this change in the function 'dropFractionalBits_fixedPointInt' where I was summing the number with a power of 2.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>