Fixed point real matrix operation
In this commit, I have coded the fixed point implementation of real
matrix multiplication. I have added two flavours of the fixed point
multiplication namely optimized and un-optimized. In the un-optimized
implementation, the accumulator has a larger bit width and the bits are
dropped only at the final stage. In the optimized version, post each
multiplication and before accumulation, the bits are dropped. Thus the
accumulator works with a smaller bit width and hence lesser area. A
point to note is that, the accuracy of the un-optimized implementation
will be slightly better than the optimized version since the accumulator
works with a larger bit width fixed point numbers and hence greater
number of fractional bits and hence greater accuracy. I have also added
a test cases file for validating the performance of the fixed point
matrix multiplication with the floating point version.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>