Function to convert a complex floating point array to fixed complex array
In this commit, I have brought a function which takes in a complex
floating point array and returns a fixed point complex array (with the
input fractional and integer bits). I could have as well used the
function which was working on real only arrays but then I would have to
separately pass the real array and imaginary arrays. To avoid 2 calls of
the same function, I have brought up this function. Another point to
note is that, python by default casts a complex value as complex 128
i.e. float 64 for real and float 64 for imaginary. Even though more
accurate, this occupies more memory. Also, mostly I will be working with
32 bit fixed point integers (for real or imaginary channels), so I have
explicitly cast the complex fixed point output as complex 64. I have
seen through test cases that the accuracy of the fixed point result with
complex 128 is about 1e-3 better than casting as complex 64. In the next
commit, I will add the test cases as well.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>