Update matrix mult testcase with complexFloat_complexFixed function call
In this commit, I have validated the complex fixed point matrix multiplication  using new function which takes in a complex floating point array and returns a complex fixed point array. I have noticed that I get better accuracy when the complex fixed point array/matrix is type cast as complex 128 instead of complex 64. This is puzzling since the values that are contained in this complex array are all well contained to 32 bit fixed point integers. So why does typecasting then matter? I need to understand this.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>