Fixed bug in generating signal of particular SNR
In this commit, I have fixed a bug. I was generating the complex
amplitude(based on the noise floor and SNR) to be applied to the signal
but I was not multiplying the signal with the phasor. Because of this,
the SNR being reported post range FFT was not correct. Just some noise
of a particular sigma was being applied. But the SNR was not being set
correctly. Now I have fixed this bug and the SNR programmed is matching
the SNR from the spectrum plot.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>