Allocate same number of bits for Twiddle factors and DFT output.
In this commit, I compute the fixed point DFT by allocating same number
of bits both the the DFT output as well as twiddle factors. i.e. If the
user specifies that he needs a 31 fractional bit DFT output, then the
twiddle factors are also set to 31 fractional bits. Similarly, if the
user wishes to have 15 fractional bits for the DFT output, then 15
fractional bits are allocated to the twiddle factors as well. I have not
verified this performance though. I will check this.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>