Compute scaled FFT using successive DFT method
In this commit, I have written a modified version of the FFT as a
successive DFT to ensure the output is scaled by N. The way to do this
is as follows. There are log2(N) stages of successive DFTs. Now, at each
stage, if we scale the ouput by a factor of 2, then the overall scaling
for the log2(N) stages becomes 2*2*....(log2(N) times) = 2**(log2(N)) =
N. Hence, if we scale the output at each stage by 2, then we can achieve
a normalized FFT.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>