Added fixed point radix 2 FFT function
In this commit, I have added the fixed point implementation of the
radix-2 FFT function. I have also added another variant of the radix-2
function which is the normalized radix-2 fft, where we divide by 2 at
each of the log2(N) stages so that the output is normalized. In the
fixed point implementation, I drop 1 LSB at each stage of the FFT. Both
the functions seem to be working fine. The normalized FFT has a higher
quantization error since we are dropping 1 bit at each stage and hence
the precision drops at each stage thus increasing the overall
quantization error as compared to the un-normalized fixed point radix 2
fft. Note that, the fixed point DFT will have a higher error as compared
to the radix-2 implementation because the DFT is un-optimal in terms of
compute ( O(N^2) computations) as against the radix 2 implementation
which has O(Nlog2(N)) computations. Infact, the exact number of real
Multiplications and Additions (MADS) in the radix 2 implementation is
5Nlog2(N). Due to this the quantization error due to fixed pointing in
DFT is much higher than the radix 2 fixed point implementation.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>