Fixed point radix-2 fft implementation
1. Changed the number of fractional bits from 31 to 15 to model a 16 bit ADC instead of a 32 bit ADC.
2. Also, the window function is quantized to a 16 bit fixed point number.
3. Disabled the DFT computation since it is time consuming and not the optimal method for fixed point FFT comparisons.
4. Implemented the fixed point radix-2 fft.
5. Also implemented the fixed point normalized radix-2 fft which scales the output by N.
6. Compared the dynamic range/noise floor set by the radix-2 normalized FFT against the theoretical value. The theoretical value for the fixed point normalized radix 2 fft is given as
N/S ratio  = 4N2^(-2b), where N/S is the noise to signal ratio, N is the number of samples of the signal, b is the number of bits for the fixed pointing of the normalized radix-2 fft. It is also the number of bits of the input signal (on which FFT needs to be done). The noise here is the total noise integrated over all the bins (or over the entire Fs). The formula can be re-written as Signal/(Total Noise) ratio = 2^(2b)/(4N). If we want the bin SNR, then we normalize the formula with N and hence the formula gets modified as bin SNR = 2^(2b)/4. Thus the bin SNR is independent of N while the total SNR is dependent on the number of samples of the signal. It makes sense because, as the number of samples increases, the number of radix-2 stages increases. This has two repurcussions. One on the signal gain and on the quantization noise floor. As N increases, the signal gain also increases. As N increases, the number of FFT stages increases leading to more FFT quantization noise added at each stage. Hence, the dynamic range/noise floor/signal to total noise ratio is dependent on the number of sampeles N.
Using this formula we can estimate the dynamic range/noise floor for a b bit fixed point FFT. This formula can be found in
Page 668 of the book Discrete-time signal processing by Oppenheim and Schafer(https://research.iaun.ac.ir/pd/naghsh/pdfs/UploadFile_2230.pdf), Eq 9.69

Now, if the ADC signal (or fixed point signal) qunatization noise floor is greater than the floor set by the fixed point fft, then the overall floor is limited by the ADC qunatization noise. But if the ADC qunatization noise is very low, then the noise floor will be limted by the fixed point FFT qunatization. I will add more informative plots in the next few commits.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>