Impose piece-wise uniform distribution on target velocities
In this commit, I have changed the velocity distribution of the targets from a Uniform(-vb-2Vs, vb+2Vs) to a piece-wise uniform distribution with -1/+1 hypothesis region 3 times as likely as -2/2 hypothesis region. Similarly, 0 hypothesis region is 2 times as likely as -1/1 hypothesis region. This is a more realistic velocity distribution of the targets, than having a uniform distribution through out from -vb-2Vs to vb+2Vs.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>