Model impact of inaccurate range update in DDMA scheme
In this commit, I have modelled and shown the impact of an inaccurate range update(per ramp) on the angle spectrum. I have used the angle error(from ground truth) and the SLL as a metric to asses the impact. Towards this modelling, I have swept the range bin offset from -2 to +2 in steps of 0.125 bins around the true range bin movement across chirps. I then plot the angle accuracy and angle spectrum SLLs as a function of the range bin offset. As expected, the DDMA scheme is very sensitive to the range update since we potentially update the range bin on a chirp by chirp basis and also phase correct per chirp(if there is a bin update). In the subsequenct commits, I will clean up the code a little more and make the plots more systematic.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>