Fixed bug in computing range bin offset sweep
In this commit, I have fixed a minor bug in computing the range bin
offset. I was computing the range bin offset sweep but was adding this
to the actual range (in meters). Ideally, I should have converted the
range and its updates to range bins and then added the range offset
sweep and then quantized the resultant. I have now fixed this bug.
Basically, the range bin offset was being added to the range instead of
range bin. Due to this, a small offset like 0.125 was being treated as
0.125m which is close to the range resolution itself and hence I was
seeing that the performance of the SLLs and accuracy was dropping very
for a small offset change. On careful observation, I found the bug and
fixed it. Now everthing is making sense.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>