Introduced Tx-Tx coupling into the DDMA model
In this commit, I have introduced inter-Tx coupling model into the DDMA scheme. When we have nearby Txs simutaneously transmitting different signals each, there could be coupling from adjacent (and other nearby) Txs thus corrupting the original signal transmitted by a particular Tx. This coupling has both a magnitude component and a phase component. Let us consider a simple signgle IC with 4 Txs say, Tx0, Tx1, Tx2, Tx3. Based on measurements results, typically, the coupling from adjacent Txs i.e Tx1 onto Tx0 (and similarly from Tx2 onto Tx1, Tx3 onto Tx2) is about 20 dB and from Tx2 to Tx0 (similarly from Tx3 to Tx1) is a further 6 dB lower. In other words, when you have all 4 Txs, i.e. Tx0, Tx1, Tx2, Tx3 all ON simultaneously and transmitting different signals each, in addition to the signal transmitted by Tx0, the signal from Tx1 couples onto Tx0 and is 20 dB lower in power. Similarly, the signal from Tx2 couples onto Tx0(in power) and is 20 + 6 dB lower, signal from Tx3 couples onto Tx0 and is 20 + 6 + 6 dB lower and so on. Roughly drops by a further 6 dB (or even lower) there onwards.  Hence the coupling from Tx1 to Tx0 on the linear voltage scale is 0.1, from Tx2 to Tx0 is 0.05 (on linear voltage scale), from Tx3 to Tx0 is 0.025 and so on. This is the magitude/amplitude coupling. On top of this, we could also have a random phase contribution from neighbouring Txs. This can be captured as a caliberation in the DDMA mode and can be applied at the receiver end to remove this effect. I have observed that the inter Tx coupling affects the SLLs in the angle spectrum. To be more particular, with the dB coupling numbers mentioned (20, 26, 32, ..), it is actually the un-compensated coupled random phase plays a bigger role in setting the SLLs than the magnitude coupling. So these random phases need to be calibeated out.

This is the Tx coupling model I have introduced. This plays a very important role in DDMA schemes. There are other factors which play a crucial role in the DDMA scheme like:
1. Non-linearity of the phase response of the phase LUT
2. Non-linearity of the magnitude response of the phase LUT
3. Bin shift
4. Phase shifter cascaded coupling.
and so on. I will try to add these models into the DDMA scheme one by one.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>