Compute and plot SLLs for DDMA montecarlo
The DDMA_monteCarlo.py script was only computing the angle
error/accuracy to characterize and verify the DDMA scheme. But that is
not sufficient. We also need to verify the Side slobe levels in order to
get a true picture of the performance of the scheme. So in this commit,
I have introduced the computation of SLLs (in addition to the angle
error). I plot the max SLL, median SLL, 80 percentile SLLs for different
SNRs. I have made 2 other changes in this commit.
1. I have made the SNR step size to 4 instead of 2 to save on the total
run time.
2. I have fixed the number of ramps to Monte Carlo for the Single IC
configuration to only 128 since that will be the most likley use case
and I can save on some compute time.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>