Put single IC Tx/Rx co-ordinates in git
The Tx/Rx co-ordinates for the single IC were not in git! Tis was a
major bug. I was previously operating from the Google Drive repository
and hence did not notice this. Now Im operating from the D drive repo.
So when I ran the DDMA_montecarlo script, immediately it threw an error
saying that the Tx/Rx co-ordinates for single IC are not available. Now
I have added these files to the git repo. Also, in this commit, I have
made another small update. I'm printing the base band velocities when
debugging for angle errors.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>