Model inter-Rx coupling in the DDMA scheme
In this commit, I have modeled the inter-Rx coupling in the DDMA scheme. This is in addition to the inter-Tx coupling that was modelled earlier. So now, the DDMA_monteCarlo.py script models both the inter-Tx as well as inter-Rx coupling. The coupling/isolation magnitude matrix is same for both Tx coupling as well as Rx coupling i.e the coupling from Tx1 to Tx0 is the same as the coupling from Rx1 to Tx0. But the random phase coupling between Txs and Rxs is different. So in the script, I use the same magnitude coupling matrix for Txs as well as Rxs. But for the phase coupling, I generate another instance of random phases for the Rxs.
Some of the other minor changes added in this script are:
1. Handle the case of a single local maxima using a try except when computing the SLL in the angle spectrum
2. Enabled the flag for coupling by default.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>