Introduce separate variable names for Tx and Rx coupling
In this commit, I have introduced separate variable names for the
inter-tx coupling and inter-rx coupling. This is more for convenience
than anything else. Also, changed the name of the flag for
enabling/disabling the Tx coupling from 'flagEnableTxCoupling' to 'flagEnableCoupling' to make it generic since the code now now caters to both inter-Tx coupling and inter-Rx coupling.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>