Model the antenna coupling in addition to the IC coupling
In the previous commits, I had modelled the IC level coupling for the Txs and Rxs. 20 dB between adjacent Txs and Rxs and a 6 dB drop from there on. In this commit, I have also added the coupling introduced by the antennas as well. So in essence there are 2 coupling factors:
1. IC level coupling
2. Antenna coupling

The typical antenna level coupling is about 15 dB if the separation is lambda/2 and falls to about 24 dB when the separation is 2*lambda. So the Tx antennas separated by lambda/2 and adjacent to each other have a coupling of 15 dB. The Rx antennas which are separated by 2*lambda and adjacent to each other have a coupling of 24 dB. I have assumed a drop of 6 dB as we move away from the Tx antennas or the Rx antennas. I have assumed a random phase coupling coupling for the antenna coupling matrix as well. But instead of varying the phase each Monte Carlo run(like I do for the IC coupling), for the antenna coupling, I initialize the magnitude and phase only once at the beginning. I dont vary this frame to frame. In terms of modelling, the antenna coupling is also a matrix (for Txs antennas and Rx antennas) which is multipled with the IC coupling matrix. So the transmitted signal is TxAntennaCouplingMatrix x TxICCouplingMatrix x Tx phase coded signal. Similary, the received signal is RxAntennaCouplingMatrix x RxICCouplingMatrix x Rx signls. For the received signal, I have introduced the coupling before adding noise. I might need to introduce the coupling after adding noise.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>